docker-api/code/requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.10.2+cpu
torchvision==0.11.3+cpu
torchaudio==0.10.2+cpu 
requests==2.26.0
fastapi==0.70.0
uvicorn==0.15.0
numpy==1.21.4
transformers==4.15.0
sentence-transformers==2.1.0
openai==0.13.0
matplotlib==3.5.1
plotly==5.5.0
azure-storage-blob==12.9.0
python-dotenv==0.19.2